add shiro, session

+++ src/main/java/script/order/DeleteOrder.java
@@ -1,6 +1,7 @@
 package script.order;
 
 import controller.FrontCommand;
+import security.AppSession;
 import service.OrderService;
 
 import javax.servlet.ServletException;
@@ -23,8 +24,16 @@ public class DeleteOrder extends FrontCommand {
 	 */
 	@Override
 	public void process() throws ServletException, IOException {
-		String orderNum = request.getParameter("orderNum").toString();
-		orderService.deleteOrder(orderNum);
-		forward("/FrontServlet?module=order&command=ViewMyOrder");
+		if (AppSession.isAuthenticated()) {
+			if (AppSession.hasRole(AppSession.USER_ROLE)) {
+				String orderNum = request.getParameter("orderNum").toString();
+				orderService.deleteOrder(orderNum);
+				forward("/FrontServlet?module=order&command=ViewMyOrder");
+			} else {
+				response.sendError(403);
+			}
+		} else {
+			response.sendRedirect("login.jsp?responseMsg=userIsNotLogin");
+		}
 	}
 }
